Crate git_nomad

Modules

git_binary ğŸ”’

See GitBinary for the primary entry point.

git_ref ğŸ”’

See GitRef for the primary entry point.

renderer ğŸ”’

snapshot ğŸ”’

types ğŸ”’

verbosity ğŸ”’

Helpers for executing Commands and parsing their Output.

workflow ğŸ”’

High level user invoked workflows for nomad.

Constants

BUILD_VERSION ğŸ”’

CONFIG_HOST ğŸ”’

CONFIG_USER ğŸ”’

DEFAULT_REMOTE ğŸ”’

ENV_HOST ğŸ”’

ENV_REMOTE ğŸ”’

ENV_USER ğŸ”’

GIT_VERSION ğŸ”’

_CARGO_VERSION ğŸ”’

Functions

build_cli ğŸ”’

Use clap to define the intended command line interface.

cli ğŸ”’

Use clap to implement the intended command line interface.

main ğŸ”’

maybe_apply_default ğŸ”’

nomad ğŸ”’

resolve ğŸ”’

Extract user arguments in order of preference:

specified_git ğŸ”’

The GitBinary intended by the user via the CLI.

specified_verbosity ğŸ”’

The Verbosity intended by the user via the CLI.

specified_workflow ğŸ”’

The nomad workflow the user intends to execute via the CLI.

version ğŸ”’