git_nomad
Module verbosity

Helpers for executing Commands and parsing their Output.

Structs

Verbosity

Responsible for timely communication of program state to the user.

Enums

CommandVerbosity

How much output to display about invoked commands.

SignificanceVerbosity

What commands to display during workflow execution.

Functions

dump_command_failure 🔒

Make some effort to build a decent error message for commands that fail.

is_output_allowed

output_stdout

Extract the printed stdout from the Output of a Command.

run_notable

run_silent 🔒

Invoke a Command and check its exit code for success.

run_spinner 🔒

run_trivial

run_with_invocation 🔒

run_with_invocation_and_output 🔒